Updated circle-rectangle collisions.
Now the lines created to test for side vs top-bottom collisions sprout at 45 degrees from the corner. Previously, they connected to the center of the plane, but this became a problem for arbitrarily long planes. A corner is the same whether it's attached to a long rectangle or not, so now all corners are treated the same.