Zombird undead mode upgrade
Previously, Zombirds in undead mode would act just like regular birds and would reflect the ball. Now, because they are ghosts, the ball will go through them when it hits them for the second time.

The major code change is that objects have the option to return false in their code() method to abort the collision. This means there can now be "ghost" collision objects everywhere, say platforms that rotate but disappear every so often. Fun stuff like that!